Note- and SpeakBrick now use FormulaEditor. Substraction a NUMBER from a Stringfunction caused and uncaught exception. Prototypeview for Mathfunctions. Delete of unused resources in strings.xml
Note- and SpeakBrick test update:

* testSoundBrickDefaultValues update
* testComputeDialog update

Fix of lint and checkstyle warnings.
Fix of Checkstyle, PMD and FindBugs warnings.

Update of Formula Editor Keyboard:

* Remove of RAND-button and update of tests.
* Update of STRING-button width.
* Math changed Functions on english keyboard.

Fix of FindBugs warnings.

+++ catroid/src/org/catrobat/catroid/content/actions/ChangeBrightnessByNAction.java
@@ -36,7 +36,8 @@ public class ChangeBrightnessByNAction extends TemporalAction {
 	protected void update(float percent) {
 		Float newChangeBrightness;
 		try {
-			newChangeBrightness = changeBrightness == null ? 0f : changeBrightness.interpretFloat(sprite);
+			newChangeBrightness = changeBrightness == null ? Float.valueOf(0f) : changeBrightness
+					.interpretFloat(sprite);
 			if (newChangeBrightness.isNaN()) {
 				return;
 			}
