+++ catroid/src/org/catrobat/catroid/content/actions/SetSizeToAction.java
@@ -36,7 +36,7 @@ public class SetSizeToAction extends TemporalAction {
 	protected void update(float delta) {
 		Float newSize;
 		try {
-			newSize = size == null ? 0f : size.interpretFloat(sprite);
+			newSize = size == null ? Float.valueOf(0f) : size.interpretFloat(sprite);
 			if (newSize.isNaN()) {
 				return;
 			}
